+++ packages/typespec-go/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Fixed handling of `multipart/form` operations.
 * Fake servers will correctly dispatch calls to all child servers, not just immediate ones.
+* Client-side default values for extensible enums that aren't in the predefined set no longer causes a crash.
+* Fixed incorrect codegen for operations that return scalar types as `text/plain`.
 
 ### Other Changes